[NO JIRA]: Migrate overlay pages
