+++ docs/src/pages/components/TouchableNativeFeedback/NativeTouchableNativeFeedback.mdx
@@ -1,0 +1,26 @@
+---
+title: Touchable native feedback
+platform: native
+githubPath: bpk-component-touchable-native-feedback
+---
+
+import Readme from 'components/Readme';
+import AndroidComponentScreenshots from 'components/Screenshots/AndroidComponentScreenshots';
+
+import readme from 'backpack-react-native/lib/bpk-component-touchable-native-feedback/README.md';
+import androidTouchableNativeFeedbackDefault from 'backpack-react-native/lib/bpk-component-touchable-native-feedback/screenshots/default.png';
+import androidTouchableNativeFeedbackDefaultDm from 'backpack-react-native/lib/bpk-component-touchable-native-feedback/screenshots/default_dm.png';
+
+
+## Table of contents
+
+## Touchable native feedback
+
+<AndroidComponentScreenshots
+  lightMode={androidTouchableNativeFeedbackDefault}
+  darkMode={androidTouchableNativeFeedbackDefaultDm}
+  altText="Android touchable native feedback component" />
+
+## Implementation
+
+<Readme file={readme} />
