+++ docs/src/pages/components/TouchableOverlay/NativeTouchableOverlay.mdx
@@ -1,0 +1,26 @@
+---
+title: Touchable overlay
+platform: native
+githubPath: bpk-component-touchable-overlay
+---
+
+import Readme from 'components/Readme';
+import IOSComponentScreenshots from 'components/Screenshots/IOSComponentScreenshots';
+
+import readme from 'backpack-react-native/lib/bpk-component-touchable-overlay/README.md';
+import iosScreenshotDefault from 'backpack-react-native/lib/bpk-component-touchable-overlay/screenshots/default.png';
+import iosScreenshotDefaultDm from 'backpack-react-native/lib/bpk-component-touchable-overlay/screenshots/default_dm.png';
+import iosScreenshotTouchableHighlight from 'backpack-react-native/lib/bpk-component-touchable-overlay/screenshots/touchable-highlight.png';
+
+## Table of contents
+
+## Touchable overlay
+
+<IOSComponentScreenshots
+  lightMode={iosScreenshotDefault}
+  darkMode={iosScreenshotDefaultDm}
+  altText="iOS touchable overlay component" />
+
+## Implementation
+
+<Readme file={readme} />
